requirements.txt
chroma-hnswlib==0.7.3
chromadb==0.3.29
huggingface-hub==0.21.3
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.35
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.14
langdetect==1.0.9
langsmith==0.1.26
numpy==1.26.4
openai==1.13.3
opencv-python==4.9.0.80
pandas==2.2.1
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pikepdf==8.13.0
pillow==10.3.0
pillow_heif==0.15.0
pyarrow==15.0.0
pydantic==1.10.14
pypdf==4.0.2
pypdf2==3.0.1
pypdfium2==4.27.0
pytesseract==0.3.10
python-dotenv
PyYAML
rank-bm25
regex
requests
rich
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.4.0
SQLAlchemy==2.0.27
streamlit==1.31.1
sympy==1.12
tiktoken==0.5.2
tokenizers==0.15.2
toml
torch==2.2.1
torchvision==0.17.1
tqdm
transformers==4.38.2
unstructured==0.12.4
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
watchdog
watchfiles
wikipedia